requirements.txt
simplejson==3.8.2
backports.ssl-match-hostname==3.4.0.2
certifi==14.5.14
Django==1.9.6
django-allauth==0.25.2
django-extensions==1.6.7
django-filter==0.10.0
django-filters==0.1.0
djangorestframework==3.3.3
djangorestframework-filters==0.3.0
django-webpack-loader==0.3.0
facebook-sdk==0.4.0
geopy==1.9.1
oauthlib==0.7.2
Pillow==3.2.0
psycopg2==2.6
PyJWT==0.4.1
python-openid==2.2.5
pytz==2015.4
requests==2.5.3
requests-oauthlib==0.4.2
six==1.9.0
wheel==0.24.0
